requirements.txt
# FastAPI Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.23

# Validation & Settings
pydantic>=2.7.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Security (Core GrabScore Template)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=41.0.0

# File Upload
python-multipart>=0.0.9

# HTTP Client
httpx>=0.27.0

# Redis (for caching)
redis>=5.0.1

# MCP (Microservice Communication Protocol)
mcp>=1.0.0

# AI Services
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.8.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
